Move clock scheduler tests into test_heroku.py
